Code clean up
Former-commit-id: ccc29bff1bbab2dbde586b76d1730c45b76ca7c6 [formerly 19f83f688542f35cd0361457e94c7e93b969c195] [formerly 6b2ef525c6c80f3d156e0211b9d453d263a877dc [formerly 7df5725eca8632f44788194fd58dc5caeed86bd6] [formerly fc047476ee40fb8001be84494e8e25e32b027aa2] [formerly 9cc9610126d66a30988cf5002f1aaffd09cdc7d4 [formerly 2298501fcacc36ea46114e65456cbe14c2aad3b4]] [formerly 6b2ef525c6c80f3d156e0211b9d453d263a877dc [formerly 7df5725eca8632f44788194fd58dc5caeed86bd6] [formerly fc047476ee40fb8001be84494e8e25e32b027aa2] [formerly 9cc9610126d66a30988cf5002f1aaffd09cdc7d4 [formerly 2298501fcacc36ea46114e65456cbe14c2aad3b4]] [formerly 2e1e62d18937f9f1c1635fa085ea6a96e49ae21b [formerly 2b6c382eb1c80facd83f7e037b914033a839f15a]]]]
Former-commit-id: 06ab8d91a7dc7fdd9018fa8d4ede6f1937ec186c
Former-commit-id: 310d2eeef11105119e39e9cbffe261291807f344
Former-commit-id: cfbe7fce68d23990077badf8a56ff650619fd52f
Former-commit-id: 3007d94af549b59a74d34061012e49bd903fe73b
Former-commit-id: cfd7dc67067b00412bda5f75f33e4d027f3dae27 [formerly caaf3463e24d95991737907da7a16d4d02705afc]
Former-commit-id: da20bd2a1816e154cd13d83fe9bf1508cf9dad12
Former-commit-id: 0caa440f6b7659d22a89e5475062991480b62406

+++ Source/RawParser/Parser/Decoder/OrfDecoder.cs
@@ -296,10 +296,10 @@ namespace RawNet
                 Tag img_entry = ifd.GetEntryRecursive(TagType.OLYMPUSIMAGEPROCESSING);
                 if (img_entry != null)
                 {
-                    uint offset = (uint)(img_entry.GetUInt(0) -img_entry.parent_offset+12);
+                    uint offset = (uint)(img_entry.GetUInt(0) - img_entry.parent_offset + 12);
                     try
                     {
-                        IFD image_processing = new IFD(reader, offset, ifd.endian,ifd.Depth);
+                        IFD image_processing = new IFD(reader, offset, ifd.endian, ifd.Depth);
                         Tag wb = image_processing.GetEntry((TagType)0x0100);
                         // Get the WB
                         if (wb != null)
@@ -339,7 +339,7 @@ namespace RawNet
                                         rawImage.blackLevelSeparate[i] = blackEntry.GetShort(2);
                                 }
                                 // Adjust whitelevel based on the read black (we assume the dynamic range is the same)
-                                rawImage.whitePoint -= (uint)(rawImage.blackLevel - rawImage.blackLevelSeparate[0]);
+                                rawImage.whitePoint -= rawImage.blackLevel - rawImage.blackLevelSeparate[0];
                             }
                         }
                     }
